Create abstract command class and finish MoreOOP
