+++ src/main/java/duke/command/ByeCommand.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.Ui;
+
+public class ByeCommand extends Command {
+
+    @Override
+    public void execute(Ui ui) {
+    }
+
+    public ByeCommand() {
+    }
+}
